+++ scribo_ct/include/scribo.h
@@ -269,7 +269,7 @@
 #       define SCRIBO__AUTO_FILL_METHOD_1  ""
 #       define SCRIBO__AUTO_FILL_TRACE_1   ""
 #   else
-#       define SCRIBO__AUTO_FILL_FATAL_1   "%s", "Game over!"
+#       define SCRIBO__AUTO_FILL_FATAL_1   "%s", "GAME OVER!"
 #       define SCRIBO__AUTO_FILL_ERROR_1   "%s", "D'oh!"
 #       define SCRIBO__AUTO_FILL_WARNING_1 "%s", "Oops"
 #       define SCRIBO__AUTO_FILL_LOG_1     "%s", "Fiat lux"
@@ -372,17 +372,28 @@
     // Implement scribo message newline
 #   if SCRIBO_SUPPRESS_NEWLINE != 1
 #       define SCRIBO__NEWLINE "\n"
+#       define SCRIBO__APPEND_NEWLINE 1
 #   else
 #       define SCRIBO__NEWLINE
+#       define SCRIBO__APPEND_NEWLINE 0
 #   endif
     
     // Implement scribo message outputting callback and flushing
 #   ifdef SCRIBO_INVOKE_CALLBACK
-        extern void scribo__ouput_message(void (*callback)(const char*), size_t size, const char* format, ...);
+        extern void scribo__ouput_message(
+            void (*callback)(const char*),
+            size_t size,
+            int newline,
+            const char* format,
+            ...);
 #       ifndef SCRIBO_SET_MAX_LENGTH
 #           define SCRIBO_SET_MAX_LENGTH 0
 #       endif
-#       define SCRIBO__DO_OUTPUT(...) scribo__ouput_message(SCRIBO_INVOKE_CALLBACK, SCRIBO_SET_MAX_LENGTH, __VA_ARGS__);
+#       define SCRIBO__DO_OUTPUT(...) scribo__ouput_message( \
+            SCRIBO_INVOKE_CALLBACK, \
+            SCRIBO_SET_MAX_LENGTH, \
+            SCRIBO__APPEND_NEWLINE, \
+            __VA_ARGS__);
 #       define SCRIBO__DO_FLUSH
 #   else
 #       define SCRIBO__DO_OUTPUT printf
@@ -402,8 +413,11 @@
             SCRIBO__DECLARE_NOW; \
             SCRIBO__GET_RAW; \
             SCRIBO__GET_NOW; \
-            SCRIBO__DO_OUTPUT(SCRIBO__TIMESTAMP_FORMAT SCRIBO__COUNTER_FORMAT SCRIBO__CATEGORY_FORMAT SCRIBO__VERBOSITY_FORMAT SCRIBO__SEPARATOR FORMAT "%s" SCRIBO__NEWLINE, \
-                              SCRIBO__TIMESTAMP_VALUE  SCRIBO__COUNTER_VALUE          CATEGORY,               VERBOSITY,                         __VA_ARGS__); \
+            SCRIBO__DO_OUTPUT( \
+                SCRIBO__TIMESTAMP_FORMAT SCRIBO__COUNTER_FORMAT SCRIBO__CATEGORY_FORMAT SCRIBO__VERBOSITY_FORMAT \
+                SCRIBO__SEPARATOR FORMAT "%s" SCRIBO__NEWLINE, \
+                SCRIBO__TIMESTAMP_VALUE  SCRIBO__COUNTER_VALUE          CATEGORY,               VERBOSITY, \
+                __VA_ARGS__); \
             SCRIBO__DO_FLUSH; \
             SCRIBO__INCREMENT_COUNTER; \
         } while (0)